chore: move metrics tracker into shared reporting
Relocate createMetricsTracker from src/plugin/src/reporting/metrics-tracker.js to src/shared/reporting/metrics-tracker.js so the instrumentation utility lives alongside other shared helpers and can be reused outside the plugin.
